src/models/rl/requirements.txt
# Core dependencies
torch>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
tqdm>=4.65.0
scikit-learn>=1.3.0

# Logging (optional)
mlflow>=2.5.0
wandb>=0.15.0

# Development dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.1.0

# Data processing (if needed)
pandas>=2.0.0
scipy>=1.11.0
